Fix admin authentication loop, synchronize API entities, and update frontend dashboard title

+++ admin-backend/src/main/java/kyung/kung_backend/domain/expert/dto/ExpertProfileCreateRequest.java
@@ -7,7 +7,7 @@ public class ExpertProfileCreateRequest {
 
     private String displayName;
     private String introduction;
-    private Long careerYears;
+    private Double careerYears;
     private Long mainCategoryId;
     private Long mainLocationId;
 }
