+++ admin-backend/src/main/java/kyung/kung_backend/domain/expert/dto/ExpertProfileUpdateRequest.java
@@ -7,7 +7,7 @@ public class ExpertProfileUpdateRequest {
 
     private String displayName;
     private String introduction;
-    private Long careerYears;
+    private Double careerYears;
     private Long mainCategoryId;
     private Long mainLocationId;
 }
